Initialize the adapter earlier to prevent an NPE
We were setting the adapter after we set the text in the
RecipientEditTextView, so depending on when the background task
ran to resolve the addresses into chips, the adapter may have been
null.

Now, we're setting the adapter as early as possible (when we have an
account), which is before we populate it.

Bug: 9408464
Change-Id: I5289cb6c36b10461167231cfc456cc018316f77b