Fix some crashes in Email
In Gmail, this change has no effect, since we currently have no
nested folders.

However, it's wrong for the teaser to take other teasers' positions
into account for positioning. The AnimatedAdapter already does this.

We were getting up in situations where the teaser was trying to be
positioned below the bottom of the list, which was causing strange
errors.

Bug: 10016802
Change-Id: I136be2367413cb7abd2724a46d448e0950ad4809