fix 2x1 attachment grids not parallaxing
For 2x1 previews, make sure to pick more a appropriate aspect ratio
that's taller than for 1x1. The dimensions are based on that tile's
drawable bounds.

Because the requests are now specific to tile bounds, add the new
variable to Request objects so requests and cached data don't get mixed
up.

Change-Id: Ide447c5eb07ddd61bf85e77b235ac1c9ccd2eced