Fix nit 5. Do not decode in background thread if cache is not warm.
When the ReusableBitmap pool is initially empty, all decodes must construct new
ReusableBitmap objects, which often causes GCs that cause jank during scroll.
This prevents the jank by notifying the ReusableBitmap pool that a scroll is in
progress, which causes the pool to block on poll() if no ReusableBitmaps can be
returned.

Bug: 10083052
Change-Id: Iaf8b244617a05d4736df041e888c9140d182c3a3