add GIF fallback support
BitmapRegionDecoder doesn't support GIFs. Upon decode failure, fall back
to BitmapFactory (and disable inBitmap as it's too picky to be useful).
This means GIF decode always allocates a Bitmap!

Even if decode fails, we should offer any inBitmap to the pool for
reuse. This was previously causing excess allocations upon GIF failure.

Because GIF decode doesn't crop to our desired parallaxable rect, amend
the rect calculation to narrow the srcRect when necessary to ensure that
some height is reserved for parallax.

Partition the pooled cache and reserve some space for non-poolable, non-
reusable entries. This is basically reserved for GIFs. Let's assume 10%
for now.

Change-Id: Iac316a2f8c62be74cca487a02d9468ede29ea787