improve senders display of the form "last, first <email>"
This doesn't exactly fix the root cause of spurious line-wrapping, but
it should reduce its incidence. We were displaying the <email> portion
of an address when the name contained a comma.

Turns out the Rfc822Tokenizer is built for comma-separated lists of
senders, with some logic to handle a sender which may itself have a
comma. Let's use that instead of TextUtils.split().

Bug: 9594770
Change-Id: I2812fb560b7395fb87f4215b211d5e824aa23083