Make some changes so the ad leave-behind can animate
First, rename INVISIBLE to something so we aren't hiding
View.INVISIBLE.

Second, create a static method for the fade-in animation of the text
so that the ad teaser can use the same code.

Third, modify CIV so that we don't need to be contained in a SCIV for
animations to work.

Change-Id: Ib89015529871bc2b7eba191dd88a4185948927ea