Fix b/10407366.
We were occasionally using the wrong
height of the html document to determine where to place
the bottom footer. We're now using the proper place for the
bottom of the webview. However, there are still some
minor issues (a little bit of content being cut off
due to some other improper use of offsetHeight).
The fix for that issue will be tracked in b/10410912.

Change-Id: I2f6dbd46f52db3809e14ddb3e3257ebbc0fbe0b4