tighten margins in conversation list items with labels
We last enlarged the vertical margin between the snippet and labels to
account for the new spacing of the star asset, but while the spec called
for 10dp between snippet baseline and labels, it was actually 14dp due
to the snippet *descender*.

Fix the excess space bug, and tighten the spacing just a tiny bit more
to 7.5dp to bring it closer to UR9's spacing.

Remove excess right-margin on the snippet. It came from a time when the
star asset had no built-in padding.

Bug: 10229604
Change-Id: Ieceeb2be34b62aa0d409aca217beb127c7e77598