analytics: avoid account_type switch on notification creation
It's nice to have notifications tagged by account_type, but because this
is a user-scope variable, it globally switches all of the user's future
UI actions to this account_type until next activity launch. That's bad.

The upshot is that currently, it's invalid to try to segment
notification and attachment-sync events by account_type. All other UI
actions are valid to segment this way.

Change-Id: Ib6c317939a8db4022cc27f64b2580469a5984a66