Improve display of attachments in pop
b/10714298
Prior to this CL, attachments that had not yet
been downloaded for POP would appear as "text/plain",
regardless of what type they really were. Also, the
filename displayed about always be a non localized
string like "click here to download".
Now, attachments in this state are flagged as dummy
attachments, so when we display them, we display a
localized instruction for how to download, and hide
the mime type (because we don't yet know it.)
Also, when we download, we will not automatically
try to launch a viewer for these kinds of attachments.
We'll just download them, and replace the old dummy
attachment with a valid one.
There is still a bug that even though the dummy
attachment has been removed from the database, we
still display that tile until the user leaves and
reenters the screen. That's ugly but not so broken
as it was before.

Change-Id: I1021baae659ca2d9ee2909f78b414c271d8ca90b