Printing: support tablet message header overflow.
b/10712542
Adds new behavior for v19 devices, both phone and tablet.
On phone, new printing option in MessageHeaderView overflow menu.
On tablet, new printing option in same place but since we now
have an overflow, we use the reply/reply all setting to show
only one of reply/reply all on tablet as well.

As a result, we also changed the string for that setting for
v19 tablets.

Change-Id: I71db88ce4a50479b03775b9d50d65e3626231589