Ensure we perform the action when we can't auto-advance
If auto-advance was set to "newer" or "older", and you archived all
your mail, the final operation would be set as the next operation to
perform in the list, but it would never actually get performed.

If we _always_ perform the action when a delete is performed, we fix
this problem.

Bug: 9904716
Change-Id: If1f0ef38a1468defabf81b385839c89334fedbc1