Shift label spaces to right of labels in RTL mode.
Fixes one part of b/11971620. Previously, when leaving
space before the label (so that multiple labels would
have a gap between them), we would always assume before
meant to the left of the label. Now, we check if we're in
RTL and if so, draw the space to the right of the label.
This has the noticeable effect of not having a gap between
the left edge of the header and the left edge of the first
label when in RTL mode.

There's still a bug in ConversationViewHeader where the
layout direction of the view is getting resolved to LTR
even when in RTL mode. Need to do more investigation.

Change-Id: Ic3a62a0678c905fe18aaf165bc7b1bb3d7631f4e