Mail in Sent, Drafts and Outbox should show addressees, not sender.
b/10847599

This is the second attempt at fixing this bug. The strategy has changed entirely to accommodate
GMail as well as reuse formatting rules that squish the list of conversation participants into an
abbreviated line for display in conversation lists.

ConversationInfo used to include a List<MessageInfo> which was used to answer questions about
which senders had read which messages in the thread. This has been removed and replaced with a
List<ParticipantInfo>. The backend should populate that list with appropriate conversation
participants (e.g. recipients of the last message in the case of Sent, Drafts or Outbox; senders
for all other mailbox types)

Change-Id: I834223c55296bcd2509fdf544634c155594739b0