Keep on getting a couldn't sign-in notification
b/11551107

This is caused by ImapConnection.doLogin() only throwing
AuthenticationFailedExceptions and not other varieties of exceptions.

Our best guess based on the reported bug is that the IMAP server should
be returning "UNAVAILABLE" indicating the IMAP server is up and
responding but some resource it relies on (e.g. LDAP) is currently down.
We translate the "UNAVAILABLE" response code into a Server Error value
and display a message to the user and a Retry button.

While fixing this bug, I discovered that the ConversationListFooterView
that is displayed in response to a authentication failure contains a
button called "Sign In" that didn't actually do anything. I made it
navigate to the incoming account settings fragment where the user is
free to change the relevant account credentials.

Change-Id: I2c772ecab18f3e57059eceeae01de08f1fdab4c2