Make sure that no null values are placed into the notification map.
Motorola IKVPREL1KK-3183. Their patch alludes to the fact that
they have hit a situation where there is a valid key but the value
associated with that key in the map was null. As a response, this
change will change the relationship with the ConcurrentHashMap from
an is-0 to a has-a to tightly control the interface into this class
and force that all puts go through a function that always allocates
an object before inserting into the map.

Change-Id: I1cc100773ba3a7b09edc59fee20e6cc14b5053f8