Make drawer available in Search and ConversationView
b/8737446
This is more complicated to do correctly than it seems like
it should be: When we are viewing search results, there is
actually a separate activity instance on top of the main
conversation list activity. This is true for both one
and two pane layouts.
So now, when we launch the search activity, we do it with
startActivityForResult. If the result is set upon return,
then the parent activity will know to navigate to a different
folder or account.

Change-Id: I5b780c4ce5bb0a1e55364cc4df280346900c5df5