Fix compose message cursor placement
Place the cursor in front of the signature when it's appended, regardless of
whether the text area is focused.

The signature gets (re)-appended in two cases:
1: When we initially set up the compose activity
and
2: When we change accounts

In neither of these cases do we generally expect the body text area to be
focused, so the code to set the cursor ahead of the signature was basically
never being run. This meant that when you tabbed into the body text area (hit
"Next" on the IME or whatever) the cursor would be placed at the end of the
text after the signature rather than, as intended, before the signature.

Thus I separated out refocusing the body text area from the cursor positioning,
and ran the cursor positioning code unconditionally when the signature changes.

b/11355420

Change-Id: I3c29ffb9734ea145eb5c6fcdc202c65669e8e2b3