Redesign of message headers. b/12933316.
Pixel perfect.
Re-architected a bunch of the layouts.
Revised some of the string building for the new formatting.
Removed unneeded references to some Views.
Fixed a visual nit in SecureConversationView with the snap header.
Removed a bunch of unnecessary BidiFormatter uses.
Deleted dead assets.
Removed padding around attachment icon and thus revised the
margin/padding for it for both conversation list and conversation
view.
Aligned star when collapsed to top of date.

Change-Id: I4b1fae35206e1f4713687c44a1c226e1257b13e4