Move Help & feedback to Nav drawer of Email.
b/11061485

Gmail remains unchanged due to this checkin.

HELP and SEND FEEDBACK are now the two bottom-most drawer
items in the Navigation drawer of the Email application
when it is present or in the overflow menu when it is not
present.

SEND FEEDBACK still behaves as always.

HELP has been redesigned so that it displays the help
content in a local webview rather than using an external
browser in the Email application.

Change-Id: If1bdba26ac2caca79c81205a409c67659cd49653