Try to use contact info from visible contacts
When querying for CP2 contact info, sort by IN_DEFAULT_DIRECTORY.
This column was recently added (earlier today), therefore only
perform this sort on android L devices.

QuickContacts won't be useful if launched with a a contact
from the default_directory (ie, an invisible contact).
QuickContacts is going to be emphasized more on android L+.
Therefore, this fix is important for L. But,
not necessarily worth fixing for previous android releases.

Bug: 13173524
Change-Id: Ib432305b2dc50ba3d67ac2a15847a98d5a93133e