Add loading view to conversation list
b/12421412
We are now supposed to show a full screen loading view,
which looks similar to the full screen empty view. This
should only display in the case where there are currently
no messages locally, but we are still trying to load them.
In this case, the full screen loading view replaces the
footer with the loading spinner.
The footer with the loading spinner still exists, for the
case when the user hits "load more" to retrieve more
messages.

Change-Id: Iba46caec2ff65c6be066ec3f9ab53bf86f4ef3e5