Commit DestructiveActions immediately. b/9904716.
For whatever reason, we have this weird two stage
commit of DestructiveActions. We call
performAndSetNextAction twice every time we wanted
to commit. This results in the next destructive action
trying to perform the action again, but that doesn't
matter because we debounce the actions so they're only
performed once.

Unfortunately, in two pane mode, when we transition back
to the conversation list, we only called performAndSetNextAction
once. As a result, the first call tried to re-apply the old
version. But it was never called a second time so we never
committed the action. Worse, the action would sit there waiting
for another action and then get applied.

Now, if we're in tablet, we set the next action so that
the single call will commit when the animation completes.

A proper fix probably involved rewriting all of this
logic and is not something I want to touch this late
in the release.

Change-Id: Iaa7b4c7d21837b93916ada5b6c9b49b13361be08