Fixed bug with % encoding
The app crashes when %25 exists in the body of mailto (actually also
crashes if it's in the subject). This is caused by attempting to
decode % encoding twice through getQueryParameters and URLDecoder.decode.

e.g. body=Bye%25Bye, after getQueryParameters. the string turns into
Bye%Bye, then passing that into URLDecoder.decode will throw error
because %Bye is not a valid % encoding.

To fix, we manually re-encode %, then run it through URLDecoder to
get the correct message.

Added unit tests to test full functionality of initFromMailTo with
various inputs to make sure that they are displayed correctly. Tested
original bug on phone, and % sign is rendered correctly with no crashes.

b/12441256

Change-Id: I06b2a4378f11de58d59b4978db1af751928570c7