Fixed draft data inconsistency
NOTE: THIS BUG APPLIES TO ALL DRAFT FIELDS (including subject,
to, cc, etc.), NOT JUST THE BODY.

The bug is caused by some kind of delayed server response where
the data inside the adapter is out of sync with the newest data.

We now force the adapter to update itself whenever any part of
the draft changes to force data consistency.

b/12587302

Change-Id: I01eea8f3aeafa14a03daad62f3462db65c3048ec