b/14075613. Don't use indices created from html on text/plain strings.
The problem resulted in a StringIndexOutOfBounds exception when composing
a reply from a plain/text only email. The tricky part is the the text entered in
the response needs to be short too. The index of the point where the quoted text
starts is built off of an html formatted buffer but applied to a text/plain buffer
in the response. This problem was not one that was seen in the Gmail app since
all mails contain html parts but the Email app is not so lucky.

Change-Id: I666c21d9066f5a2f4dc7131fa151e34ff6193ee2