Switch to default inbox if no folder is selected
Upon new account creation, we might enter the main activity after
the inbox has been synced into place, but we won't have navigated
to it yet. As a result, the selection was falling back to the first
folder in the cursor, which was the starred folder.

We can detect this case by observing that mSelectedFolderUri is empty.

b/14319452

Change-Id: I4402f42acd7c845ee33db91d3a0037a068391b85