Properly destroy old loaders in MessageFooterView
Each MessageFooterView assumes that it only gets updates from one
corresponding attachment uri for its attachment information. However,
this can be broken if we change the underlying uri in MessageHeaderItem
without informing the footer view. When the next bind operation comes
around, MessageFooterView will see the new uri, assume that it's the old
one, and deduce that the uri hasn't changed. It will then attach a
loader to the new uri, and now it has two loaders feeding information
on two different states of the attachment list.

And this will then cause (50% of the time depends on which loader
finishes last) the view to render both the 2 attachments and 1 attachment,
and depending on which one is rendered last, the one with more attachments
will either completely cover up the other, or the one with fewer
attachments will incorrectly overlay on top.

b/14293222

Change-Id: Ia652601d26ce14e9aa0346a98dad2597e62bdbaf