Fix view state when resuming from search
Added check to onResume such that when we resume and the cursor is
not ready, we should essentially treat it like onCreate and set the
appropriate flags and the default view state. Previously, the fragment's
cursor gets cleared when we perform a search, thus when we resume from
search, we see the empty state for a little bit before the cursor loads
the actual data.

b/14461247

Change-Id: I5f9124179e710631aa76b414dd1fec3c4d670138