Minimum html support using Html class (no ol/ul)
Support basic html in ComposeActivity using Html and HtmlTree class.

- Extended HtmlTree to take in any arbitrary html -> object converter
  so we can use it to convert html to spanned as well as plaintext.
- Added custom SpannedConverter to feed into HtmlTree.
- Added async task to load Utils.htmlToSpan in the background.
- Moved ComposeActivity#sendOrSaveInternal to a handler thread in
  case Utils.spanToHtml takes a long time.

Note that we only support a subset of the html tags. To see the full
list of tags we currently support, refer to HtmlUtils#SpannedConverter.

Performance:
About 3x faster than using Html.fromHtml (yay!).
Performance should not be an issue for >99% of the users.

b/7406122

Change-Id: I484d7fc7d2f36c4724d9e4cb90fa1d65eadcb903