Support for blowing up cache on subsequent search
Add support to get fresh results every single time search is called
even when the query is exactly the same.

The current behavior is that it destroys the old data once the query
changes. This normally works fine because when we switch labels the
query will be different while scrolling down in the same label ensures
that we don't lose the old data. However, for ACTION_SEARCH, the
underlying provider have no way of knowing whether the action is a
fresh search or simply the user loading more messages. Adding the
additional query parameter used by search will allow us to
differentiate between the two cases and properly destroy the old data.

b/11031046

Change-Id: I7fa244af8c13a6a39d68c846fa4a0db605b9b74a