Fixed conditional for checking if cursor is ready
Because the cursor can be mutated by user input, we should deem
the cursor to be ready to show if it either has a count of 0 or
it has a COMPLETE state. Prior we were checking if it's not in
LOADING state, which can be wrong if the cursor is partially
loaded but the user removed all of the loaded conversations. In
that case, the cursor's size would be 0 but the state would be
LOADED, and it would wrongly show the empty state.

b/15583795

Change-Id: I7c953a477922d10c027e40b19e8fe83a09b968e4