Fix bug with omitting checked system labels
Since we now no longer need to manually sort the selected labels
to top of the list since we have separate section for selected
and unselected labels, we can remove the check for isSelected
when adding system labels in their default order before calling
folderSort. This bug was causing selected system labels to not appear
under the systemLabels list AND nonsystemlabels list, therefore
it was missing completely in the dialog.

b/15777487

Change-Id: I14ba2872f87d1c9baed7e9143c8b7e8219f33c45