More robust way to cancel conversation notifications.
This happens when users archive or delete email from the watch. Currently the
conversation notification is cancelled in
NotificationUtils.validateNotifications()[line 775] by comparing current ids
with previous ids stored in memory. If the gmail app is restarted, previous
ids will be cleared and some conversation notifications may not be properly
cancelled.

The solution is to add the notification id to the notification action.
NotificationActionIntentService will use the id to cancel the notification if
it is a remote action.

Bug 16371569
Bug 16376756
Bug 16020553

Change-Id: If70e075e258823906fc062b89ce9b023b419680d