Address NFE with bgColor
We were only checking if the string was null
instead of checking if it was empty. Switched
conditions from != null to !TextUtils.isEmpty(..).

I'm leaving this bug open for now to see how
bgColor ended up being an empty string. Also, I've
added safeguards for fgColor as well to make sure
it never throws an NFE.

Bug: 17139903
Change-Id: I1263a8ba2f4aafae01af0ee72768368cdcac5e4b