Fix last email archive (No more lost archives!)
Took away the quick return after showNextConversation. It
seems to be causing the last action to not archive/delete
as there is no conversation to show, which leads to the
if registering as true and returning without finishing
the deletion.

There is still something weird with calling delete twice,
but at the least the bug is fixed (I'll continue looking
into the unnecessary delete call later at b/17290364)

Bug: 9904716
Change-Id: If0af5502dc8363a2e4267dbdcf7c79c06f15dff1