Add support for disabling email sanitization
b/16896849

HtmlSanitizer defines a public static final int VERSION that should be
bumped any time a change to the sanitizer's configuration is made.

ConversationViewPager now compares the current sanitizer version
against the target version number from MailPrefs to decide if the
sanitized output is "safe". The Html is displayed in a sandbox without
scripting if it isn't safe.

Change-Id: Ia265180893dbdd9ef209d3ee6ea87c7eb468e6b3