Refactor out SendOrSaveTask
sendOrSaveInternal is already run on a handler, so there is now
no need to have SendOrSaveTask. Change SendOrSaveCallback
accordingly since some of the args are no longer needed.

This is a speculative fix for b/17425799 since we were double
posting tasks to the same handler. As a result, there can be a
race condition where the first SAVE posts to handler, doesn't
complete in time before the second SEND posts to handler. Thus,
the SAVE's SendOrSaveTask is run after the SEND runnable.

b/17425799

Change-Id: I3a2379cb254854598c0d01aef1993e3d21d13788