Fix keyboard RTL navigation
Main issue is that we no longer need the focus changes
in onViewModeChanged because we handle left/right navigation
individually.

Furthermore, changed isRtl check to using the ViewCompat version
instead of the locale version so it's easier to test (locale
version doesn't work for force rtl in dev options).

b/17234283

Change-Id: I3f87e67ed1c7457370f8cdb63cc3abaa57211db9