Update shadows for TL (Quantum) UnifiedEmail
NOTE: I'll fix FAB issues in a different CL

Add shadows assets on L & pre-L (RTL included)
for thread list sides. To avoid overdraw, the
backgrounds have been modified (The outermost
framelayout for TL now has a white background
instead of each of the conversation items).

For both L & pre-L, we're using two shadow assets
on either side of the TL, both of which are
set as backgrounds for the drawer and
the conversation view (conv pane). Elevation
didn't work out too well and will be revisited
while fixing overdraw issues (Need to move around
swipe/leave-behind item backgrounds and it
seems a bit risky to mess with it given the
timeline)

+ Also removed the background on the loading
item under conversations in the TL and used
the default conversation item background to
get rid of the ugly divider/slight line flicker
on folder change

Bug: 17461682
Change-Id: Id9e4504b6063f21a055ed492fc8de962e6e52462