Fix navigation to mini drawer
Soemthing changed for current_account_avatar (I see it
in xml, but the id is never referenced in code and I
couldn't find it in hierarchy viewer) which broke navigation
since that was the default navigation id I was using for
mini drawer. I changed the default id to the mini drawer
itself and auto-passing the focus to its first child.

Also remove focusableInTouchMode for FAB.

b/17555177

Change-Id: I79f27850efdcc06497a042d1e5b155a6755fdbed