Gmail crashes when switching to Promotion tab
b/17534190

Recently, mHeader.styleNames usage changed.

Before: A new ArrayList was built and assigned to this field on each
pass through calculateTextsAndBitmaps.

After: It was changed to build the ArrayList once and .clear() it on
each pass.

This bug was caused by an obsolete NULL check of mHeader.styleNames
that would always be true post-change.

Change-Id: I9b794a96093027a2eee7b7c38b0452d273184968