Ad icon is misplaced in RTL format
b/17285804

BidiFormatter to the rescue. It was not being applied to the subject
text, which was then combined with fixed badge text from our resource
files. Chaos ensued.

unicodeWrap()ing the subject text was the solution

NOTE: the code continues to assume the ad ALWAYS appears at the
beginning of the combined string and thus before the subject.

Change-Id: I1aeefd0808284a4a85fe10e5ca1f8f18822c55e2