New empty states for tablet landscape
For tablet landscape, the empty icon should be shown
in the CV pane while the empty text (if the folder is
empty) should be shown in the list pane.

Added a new empty icon for not having any selected conversation,
this state does not have text associated with it since
the list pane still showed the available items.

b/16876067

Change-Id: I27b7b9ae8ce739f23829afc292df083e8399a118