Repair "New Messages" snackbar on phones
b/17373672

When code was added to enforce minimum and maximum widths for
the benefit of 2-pane snackbar layouts, and unintended consequence
was effectively hiding the "New Messages" snackbar on phones.

This is because both the "New Messages" snackbar and the real
snackbar in AAC are implemented with the same class:
ActionableToastBar.

To fix this, we detect the existence of negative min and max widths
and respond by skipping the width bounds enforcement within
ActionableToastBar. Essentially, negative values indicate the
natural width of the ActionableToastBar should be preserved.

Change-Id: If9b9b681bab2aa8b5dc6da5ed514dbcb2542a46d